refactor: share map features
